Expose methods AlphaView, BetaView from Evaluator.
This provides a means to see the matrix entries in different ways.

In particular, now it is easy to see the entries: as-is, log-spaced
with the column scales applied, or as the actual log-probabilities
(with column scales and counterweights applied).

(Meanwhile, since ToHostMatrix no longer necessarily returns
log-scaled values, remove that part of the API comment)